Make the plugin compatible with Gambio GX 4.1.1
The configuration database table is gone and one of the DTOs of the CookieConsent Plugin has a new constructor signature.
Detecting the new constructor is done by sheer brute force, but as long as the api is not stable yet it is not worth the effort of implementing that properly.

The callback and hooks should be refactored to not rely on deprecated functions (ie. gm_get_conf() and defines set based on config values) but it works for now.
We'll wait until the api is considered stable by Gambio before rewriting that.