Fix the menu for Gambio GX4.3.1.0
The xml file contains an attribute which breaks GX4.3.1.0. Older versions ignore this attribute.
GX4.2.x supports menu descriptions as json files. Add one once it is needed. It is disabled for now (one dot too many in the file name) because otherwise the menu will be added twice.
GX3 does not support .menu.json so removing menu_relevanz.xml is not an option yet.